more eureka tests updates for runAsAdmin false

+++ cypress/e2e/eureka/authorization-roles/assigning-users/assigning-users-in-role-view.cy.js
@@ -39,8 +39,7 @@ describe('Eureka', () => {
                 });
                 cy.createAuthorizationRoleApi(testData.roleBName).then((role) => {
                   testData.roleBId = role.id;
-                  // TO DO: rewrite when users will not have admin role assigned upon creation
-                  cy.deleteRolesForUserApi(testData.userA.userId);
+                  if (Cypress.env('runAsAdmin')) cy.deleteRolesForUserApi(testData.userA.userId);
                   if (Cypress.env('runAsAdmin')) cy.updateRolesForUserApi(testData.userB.userId, [testData.roleBId]);
                   else cy.addRolesToNewUserApi(testData.userB.userId, [testData.roleBId]);
                   cy.login(testData.tempUser.username, testData.tempUser.password, {
