+++ cypress/e2e/eureka/users/assigning-roles-user-edit.cy.js
@@ -71,8 +71,7 @@ describe('Eureka', () => {
 
     before('Assign roles, login', () => {
       cy.getAdminToken();
-      // TO DO: rewrite when users will not have admin role assigned upon creation
-      cy.deleteRolesForUserApi(testData.userA.userId);
+      if (Cypress.env('runAsAdmin')) cy.deleteRolesForUserApi(testData.userA.userId);
       cy.login(testData.tempUser.username, testData.tempUser.password, {
         path: TopMenu.usersPath,
         waiter: Users.waitLoading,
